Creates a confirmation dialog thats utilized by the application through a provider

+++ src/pages/Teams/Teams.jsx
@@ -1,13 +1,15 @@
 import style from "./Teams.module.css";
 import {useEffect, useState} from "react";
 import dataBase from "../../services/DataBase";
-import {useToast} from "../../Provider/ToastProvider";
+import {useToast} from "../../provider/ToastProvider";
 import {Link} from "react-router-dom";
+import {useConfirmation} from "../../provider/ConfirmationProvider";
 
 export default function Teams() {
     const [user, _] = useState(JSON.parse(localStorage.getItem("user")));
     const [teams, setTeams] = useState([]);
     const {showToast} = useToast();
+    const { showDialog } = useConfirmation();
     const CREATION_TEAM_ID = "60753bbe-2c1f-4e40-963a-21ea6c14c777";
 
     const fetchUserTeams = async () => {
@@ -23,6 +25,10 @@ export default function Teams() {
     }
 
     const handleDelete = async (teamId) => {
+        const isConfirmed = await showDialog("Are you sure you want to delete this team? This action cannot be undone.")
+
+        if (!isConfirmed) return;
+
         try {
             const isDeleted = await dataBase.deleteTeam(teamId);
             if (isDeleted) showToast("Team deleted");
@@ -34,6 +40,10 @@ export default function Teams() {
     }
 
     const handleDeleteAll = async () => {
+        const isConfirmed = await showDialog("Are you sure you want to delete all teams? This action cannot be undone.")
+
+        if (!isConfirmed) return;
+
         try {
             for (const team of teams) {
                 await dataBase.deleteTeam(team.id);
@@ -44,6 +54,7 @@ export default function Teams() {
             console.log("Erro ao deletar todos os times: ", err);
             showToast("Erro ao deletar todos os times");
         }
+
     }
 
     useEffect(() => {
@@ -53,7 +64,6 @@ export default function Teams() {
     return (
         <section className={`${style.teamSection} flex-column align-center smallPadding mediumGap`}>
             <h2>Teams</h2>
-
 
             <div className={`${style.buttonContainer} flex-row justify-center smallGap smallPadding`}>
                 <Link to={`/teams/teamEditor/${CREATION_TEAM_ID}`} className={`flex-row flex-center smallPadding`}>
@@ -100,5 +110,5 @@ export default function Teams() {
 
             <div className={`largePadding`}><br/><br/></div>
         </section>
-    );
+    )
 }
